requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
pyaudio
speechrecognition
setuptools
packaging
langchain_community
PyPDF2